Huge fix the issue with associativity and with the unraty operator order.
It is supposed to be noticed that recursive call of parse in parse_expr is the right dicision,
instead of calling parse_expr as well. Since it will be not a left-associative.
Just take a chance to mention that it is a small difference has a huge impact, as everywere and everytime in our life.
Take care.