il: 3ac: resolve issue with wrong scope variables
The context structure now looks not as straightforward as
it deserve to be.

The context structure responsible for dealing with scopes and managing ID.
And to be better showed in pretty stage it pass all allocatation
to pretty_output by FundDef structure.
That's pretty all and for this purpouses it's to complicated.

Might we can develop anather structure for managing ID