translation: x64: change logic parsing values.
There was a problem with sub operation.

After that I found out that we should manage in better way +, * etc. instructions.
since it's often when we got such sort of instructions `add %eax, %eax`, what is exactly wrong.
It happens when the lhs and rhs values stored by the same place. We should track it somehow.