translation: x64: add support of multi return statements
It checkes if there's a more then one return statement or
the one moved somewhere in loop for example and if it is it generate new label
on translation stage. The label responsible to be jumped on when
we run into tac::return statement.

It complitely differs from `clang`, `gcc` approached I as can tell,
at least since they have the same type of label, here we don't.

Generally it implemented with bad taste.
It showes sings of bad designe I assume.

The questions which I can ask here.

- TAC/3ac representaion currently is OK?
  Might there's reasons to restructure return statement.
- Does this passing information about return consistenct well in the `func_begin` function.
- Does translation layer shoud be responsible for this?