translator: move to IR -> x64 asm implemenation
This is reimplemenation of translation stage, to get deal with IR
rather than AST directly. The main accent has been put to be able to use
registers as allocation place.

The alocation algorithm was chosen is
'Linear allocation algorithm' because its simplicity and more importantly
there's almoust no absrations so its hard to amalgamate every part. Spilling is might deserve got rethinging.
I ponder how many errors there are.

The borders between alocator and translation process were washedout.
Which is a bad design I guess. There's no tests at all. I used 9 stages of Nora Sandler's test suit to verify correctness.

That was a beastly long jorney to get here. The tries and approaches were started
a month and a half ago. And that's absurd in what way it turned out eventually.
But I love it.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>