lexer: revert 1373667d8a97a968bb4c8d979413760d72dbe07d
In parsing stage we can't parse such expression like (a-1)
because of that lexing rule.

example:
```c
int n = getchar();
do_with(n-48);
```

Initially that rule was created to win in performance by not using unary operation at all.
But as we see it does not work in such way.
In any way with constant folding it must be solved as well.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>