add support of global variables
It took quiet an effort to have done.
The code equality is going down with this commit.
But I decided to keep that in that shape in order to
see problems more clearly. It displays lots of issues most of which
is code is highly relatedness. In short abstractions don't work well.

Comments:
* allocator have a notion of formating labels now
* doesn't support constant folding from AST
* functions get it's own scope, which is good :)
* new checks showed the visitor trait in descent light

I have only one believe it does not spawn more bugs :)

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>