generator: add support of intel syntax
Intel syntax require `.intel_syntax noprefix` directive,
currently it's put in `syntax` module, but by that we change the structure of
the output assembly in syntax module, which was plainly developed to avoid it.
Might it would be better to have the changes of assembly in generator code or at least
provide the clear structure of assembly function in `Assembly` structure, since potentially
the changes in generator will break the directive adding. It will be placed just in unexpected place.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>