gen: copy parameters on stack
Aparantly it's required to copy it.
As if the function has a call any register may be messed.

The implementation took place in allocator, which might not a good idea
since allocator started being knowing lots of information about a target arch.
And what's more now `Allocator::new()` returns additionaly a `asm::Block` of copy oprations.